requirements.txt
beautifulsoup4==4.6.0
lxml==3.8.0
Pillow==4.1.1
pynder==0.0.13
pyparsing==2.2.0
requests==2.18.1
robobrowser==0.5.3
urllib3==1.21.1
boto3==1.9.2
scikit-image==0.16.2
scikit-learn==0.18
protobuf
Theano==0.7.0
opencv-python
lasagne==0.1
scipy==1.1
